Fix locales
Amend Dockerfile to include locales and use the working dir `/app`

And, add logic to explicitly create the absolute path and to log where it's looking for locales, for debugging purposes. As well, added some error "handling"

Explanation:

Iniitial problem is that locales are not included into the docker image by default.

Second problem is the application was looking for loacales in it's `cwd`. But because the dockerfile workdir was `/` wiith the binary in `/app` it was looking in `/locales` instead of `/app/locales`